+++ frontend/src/modules/integration/components/platforms/integration-custom.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script setup>
-import AppIntegrationListItem from './integration-list-item'
+import AppIntegrationListItem from '../integration-list-item'
 import { defineProps } from 'vue'
 
 defineProps({
