+++ frontend/src/modules/integration/components/platforms/integration-devto-drawer.vue
