+++ frontend/src/modules/integration/components/platforms/integration-discord.vue
@@ -19,7 +19,7 @@ export default {
 }
 </script>
 <script setup>
-import AppIntegrationListItem from './integration-list-item'
+import AppIntegrationListItem from '../integration-list-item'
 import { useStore } from 'vuex'
 import { useRouter } from 'vue-router'
 import { defineProps, computed, onMounted } from 'vue'
